pyproject.toml
[tool.poetry]
name = "reverse-engineering-domain-concepts-from-code-using-llms"
version = "0.1.0"
description = ""
authors = ["IbrahimOued <brahim.oued@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
stanza = "^1.8.2"
matplotlib = "^3.9.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.0.1"
mlflow = "^2.14.1"
scikit-learn = "^1.5.0"
seaborn = "^0.13.2"
transformers = "^4.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
